Server: Example server to test the API
The server is working, you have to type it's full address in the game's
entry (e.g. `http://localhost:8888`).

`GET /boards` returns the size of the boards (and the cells when I
implement it), `POST /action` will be used for sending actions (move
ship, fire, etc.).

WARNING: Commit introducing TODOs